# Optimize AssetCollectionView