--- a/mediathread/assetmgr/models.py
+++ b/mediathread/assetmgr/models.py
@@ -1,10 +1,12 @@
+import json
+import re
+
 from courseaffils.models import Course
 from django.conf import settings
 from django.contrib.auth.models import User
+from django.core.cache import cache
 from django.db import models
 from tagging.models import Tag
-import json
-import re
 
 
 class AssetManager(models.Manager):
@@ -165,7 +167,7 @@ class Asset(models.Model):
 
     @property
     def html_source(self):
-        return Source.objects.get(asset=self, label='url')
+        return self.source_set.get(asset=self, label='url')
 
     def xmeml_source(self):
         return self.sources.get('xmeml', None)
@@ -176,19 +178,21 @@ class Asset(models.Model):
 
     @property
     def primary(self):
-        if not hasattr(self, '_primary_cache'):
-            self._primary_cache = self.source_set.get(primary=True)
-        return self._primary_cache
+        key = "%s:primary" % (self.id)
+        if key not in cache:
+            cache.set(key, self.source_set.get(primary=True))
+        return cache.get(key)
 
     @property
     def thumb_url(self):
-        if not hasattr(self, '_thumb_url'):
+        key = "%s:thumb" % (self.id)
+        if key not in cache:
             try:
-                self._thumb_url = \
-                    Source.objects.get(asset=self, label='thumb').url
+                url = self.source_set.get(label='thumb').url
             except Source.DoesNotExist:
-                self._thumb_url = None
-        return self._thumb_url
+                url = None
+            cache.set(key, url)
+        return cache.get(key)
 
     def tags(self):
         # returns all tags for this instance's notes