--- a/mediathread/settings_shared.py
+++ b/mediathread/settings_shared.py
@@ -59,6 +59,12 @@ PROJECT_APPS = ['mediathread.main',
 TEST_RUNNER = 'django.test.runner.DiscoverRunner'
 
 CACHE_BACKEND = 'locmem:///'
+CACHES = {
+    'default': {
+        'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
+        'TIMEOUT': 3600  # One Hour
+    }
+}
 
 TIME_ZONE = 'America/New_York'
 LANGUAGE_CODE = 'en-us'