--- a/mediathread/templates/test.html
+++ b/mediathread/templates/test.html
@@ -1,0 +1,4 @@
+<html>
+<body>
+</body>
+</html>