--- a/mediathread/templates/test.html
+++ b/mediathread/templates/test.html
@@ -1,4 +1,0 @@
-<html>
-<body>
-</body>
-</html>