… requirements/src/django_courseaffils-2.0.3.tar.gz