The AssetCollectionView (/api/asset/) was proving very slow in situations where items had a large number of selections. Testing against a course with 35 items (low) and 3,043 selections (high average per item)

Original metrics: 4597 queries / 4580ms

4597 queries / 4301.61ms
* use cached_course_is_faculty wherever possible
* remove is_authenticated from check. you can't get here if you're not authenticated

4302 queries / 4452.57ms
projects/models.py -- get content_object in bulk

1304 queries / 2695.24ms
cache Asset primary and thumb_url. (needs to be rearchitected)

875 queries / 2408.11ms
select_related in Collaboration.get_for_object functions

622 queries / 2239ms
use existing collaborations rather than fetching again